+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adhere's to [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 
 ## [unreleased]
 
+- Add `onSaveCommand` to `LanguageConfig`. Can now configure a command to run on saved files.
+- Restarting Alloglot no longer disposes of its main output channel.
+
 ## [3.0.4]
 
 - Fix crashing `Alloglot: Restart` command.